Fix step handling regression
When processing the steps if false is returned the item should be skipped.
However there are two problems. The test for the return of the step->process
function tests any value that evaluates to false. So MappingStep others that
return nothing cause the row to be skipped. Therefore the test now checks
that the value is fully equal to false. The second issue is that when a step
returns false we continue, but since its a nested foreach loop, it only stops
the nested loop. So it stops processing the steps but moves on to the writers.